pyproject.toml
[project]
name = "minha-crew"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "crewai-tools>=0.17.0",
    "crewai>=0.86.0",
    "pypdf2>=3.0.1",
    "streamlit>=1.41.1",
    "python-dotenv>=1.0.1",
]
